Working streaming API for moggcrypt + vorbisfile, preview now plays mogg directly

+++ haskell/packages/onyx-moggcrypt/cbits/XiphTypes.h
@@ -1,7 +1,8 @@
 #pragma once
 
 #include <stdio.h>
-typedef long long ogg_int64_t;
+#include <stdint.h>
+typedef int64_t ogg_int64_t;
 typedef struct {
 	size_t(*read_func)  (void *ptr, size_t size, size_t nmemb, void *datasource);
 	int(*seek_func)  (void *datasource, ogg_int64_t offset, int whence);
